try to clean up App.tsx a bit
